% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dilocus_allele_freq.R
\name{calcular_frecuencias_dilocus}
\alias{calcular_frecuencias_dilocus}
\title{Calcular frecuencias genotipicas y alelicas en un marcador dilocus}
\usage{
calcular_frecuencias_dilocus(tabla)
}
\arguments{
\item{tabla}{Un data frame que contiene dos columnas:
\describe{
\item{Genotipo}{Vector con los genotipos observados (por ejemplo, "AABB", "AABb", "AAbb", etc.).}
\item{Frecuencia}{Vector con las frecuencias observadas de cada genotipo.}
}
La tabla debe tener 9 filas correspondientes a los genotipos combinados para dos loci.}
}
\value{
Un lista con los siguientes elementos:
\describe{
\item{Frecuencias genotipicas}{El data frame original con la columna adicional de frecuencias genotipicas.}
\item{Frecuencia alelica locus A/a}{Un vector con las frecuencias alelicas para el locus A/a (p_A y q_A).}
\item{Frecuencia alelica locus B/b}{Un vector con las frecuencias alelicas para el locus B/b (p_B y q_B).}
\item{Frecuencia alelica Haplotipo}{Un vector con las frecuencias de los haplotipos AB, Ab, aB y ab.}
}
}
\description{
Esta funcion calcula las frecuencias genotipicas y alelicas en un marcador de dos loci (dilocus),
dada una tabla con los genotipos y sus frecuencias observadas. Ademas, calcula las frecuencias de los haplotipos combinados
para los loci involucrados.
}
\examples{
#Crear una tabla de ejemplo con genotipos y frecuencias
dilocus_tabla <- data.frame(
  Genotipo = c("AABB", "AABb", "AAbb", "AaBB", "AaBb", "Aabb", "aaBB", "aaBb", "aabb"),
  Frecuencia = c(0, 0, 0, 38, 0, 0, 2, 0, 0))

calcular_frecuencias_dilocus(dilocus_tabla)

}
