% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWQL1.R
\name{calculate_HWQ}
\alias{calculate_HWQ}
\title{Calcular el Test de Hardy-Weinberg}
\usage{
calculate_HWQ(genotype_counts)
}
\arguments{
\item{genotype_counts}{Un vector numerico que contiene los conteos de los genotipos observados.
Debe tener tres elementos, correspondientes a los genotipos \code{AA}, \code{Aa} y \code{aa} en este orden.}
}
\value{
Un lista con los siguientes elementos:
\item{Total_Individuals}{Número total de individuos observados en la muestra.}
\item{Genotypic_Frequencies}{Un data frame con las frecuencias genotípicas observadas.}
\item{Allelic_Frequencies}{Un data frame con las frecuencias alélicas observadas.}
\item{Expected_Frequencies}{Un data frame con las frecuencias esperadas según Hardy-Weinberg para los genotipos \code{AA}, \code{Aa} y \code{aa}, y sus correspondientes conteos esperados.}
\item{Observed_Counts}{Un data frame con los conteos observados de los genotipos \code{AA}, \code{Aa} y \code{aa}.}
\item{Chi_Squared}{Valor del estadístico chi-cuadrado calculado para la prueba de ajuste Hardy-Weinberg.}
\item{P_Value}{P-valor asociado al estadístico chi-cuadrado para determinar si la diferencia entre las frecuencias observadas y esperadas es significativa.}
}
\description{
Esta función realiza un test de Hardy-Weinberg utilizando los conteos de genotipos observados.
Calcula las frecuencias genotípicas y alélicas observadas y esperadas según Hardy-Weinberg,
luego evalúa si la población está en equilibrio usando el estadístico de chi-cuadrado y su p-valor.
}
\examples{
# Ejemplo con un conjunto de datos de conteos genotípicos
mi_tabla <- c(AA = 4, Aa = 33, aa = 3)
resultados_HWQ <- calculate_HWQ(mi_tabla)
print(resultados_HWQ)

# También puedes obtener los conteos observados y las frecuencias esperadas
resultados_HWQ$Expected_Frequencies
resultados_HWQ$Observed_Counts

}
