% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createArray2.r
\name{createArray2}
\alias{createArray2}
\title{create array for representing genotypes and niches}
\usage{
createArray2(...)
}
\arguments{
\item{...}{named dimensions to create in array e.g. dimname=c('name1','name2')}
}
\value{
array
}
\description{
creates array with the passed dimensions
}
\examples{
createArray2( sex = c('f','m') )
#Wloci
createArray2( loci=c('SS1','RS1','RR1','SS2','RS2','RR2'), niche1=c('0','a','A'), niche2=c('0','b','B') )
createArray2( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2') )
createArray2( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2'), niche1 = c('0','a','A') )
createArray2( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2'), niche1 = c('0','a','A'), niche2 = c('0','a','A') )
}

