% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createGametes.r
\name{createGametes}
\alias{createGametes}
\title{create gametes array}
\usage{
createGametes(f, recomb_rate)
}
\arguments{
\item{f}{array with frequencies of genotypes in the popn.}

\item{recomb_rate}{recombination rate}
}
\value{
array
}
\description{
create gametes based on the frequency of genotypes in the population and the recombination rate.
This is refactored from code in runModel()
}
\examples{
#createGametes()
}

