% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createInputMatrix.r
\name{createInputMatrix}
\alias{createInputMatrix}
\title{create input parameters matrix}
\usage{
createInputMatrix(params.csv, inFile = system.file("extdata",
  "input.parameters.csv", package = "resistance"), calibration = 1012,
  save.fitvals = 0)
}
\arguments{
\item{params.csv}{whether parameters are defined in an input file TRUE/FALSE}

\item{inFile}{parameters filename}

\item{calibration}{one of a limited set of integers effecting how scenarios are run}

\item{save.fitvals}{whether to save fitness scores to csv 0=no, 1=yes}
}
\value{
matrix
}
\description{
Creates a matrix with parameters in rows and scenarios in columns.
Either from a csv file or hardcoded scenarios set by the calibration argument.
}

