% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diloc_Freq_Hd.R
\name{Diloc_Freq_Hd}
\alias{Diloc_Freq_Hd}
\title{Calculate Allele Frequencies for a Biallelic Marker}
\usage{
Diloc_Freq_Hd(data)
}
\arguments{
\item{data}{A data frame where each row represents a population, the first column is the population name (\code{Pop}),
and the remaining columns are genotype counts (\code{AABB}, \code{AABb}, \code{AAbb}, \code{AaBB}, \code{AaBb}, \code{Aabb}, \code{aaBB}, \code{aaBb}, \code{aabb}).}
}
\value{
A data frame with haplotype frequencies (\code{AB}, \code{Ab}, \code{aB}, \code{ab}) for each population.
}
\description{
This function calculates haplotype frequencies for a biallelic genetic marker based on a genotype distribution
for multiple populations. The input is a data frame with the genotype counts for each population.
}
\examples{
# Create a data frame with genotype counts for three populations
DicGtps <- data.frame(
  Pop = c("Pop 1", "Pop 2", "Pop 3"),
  AABB = c(5, 3, 4),
  AABb = c(3, 2, 5),
  AAbb = c(4, 2, 5),
  AaBB = c(4, 6, 3),
  AaBb = c(10, 8, 12),
  Aabb = c(6, 5, 7),
  aaBB = c(3, 4, 2),
  aaBb = c(2, 1, 3),
  aabb = c(3, 2, 4)
)

# Calculate haplotype frequencies for the data frame
Diloc_Freq_Hd(DicGtps)

}
