% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Freq_Diloc.R
\name{Freq_All_Diloc}
\alias{Freq_All_Diloc}
\title{Calcular frecuencias genotipicas y alelicas en un marcador dilocus}
\usage{
Freq_All_Diloc(tabla)
}
\arguments{
\item{tabla}{Un data frame que contiene dos columnas:
\describe{
\item{Genotipo}{Vector con los genotipos observados (por ejemplo, "AABB", "AABb", "AAbb", etc.).}
\item{Frecuencia}{Vector con las frecuencias observadas de cada genotipo.}
}
La tabla debe tener 9 filas correspondientes a los genotipos combinados para dos loci.}
}
\value{
Un lista con los siguientes elementos:
\describe{
\item{Frecuencias genotipicas}{El data frame original con la columna adicional de frecuencias genotipicas.}
\item{Frecuencia alelica locus A/a}{Un vector con las frecuencias alelicas para el locus A/a (p_A y q_A).}
\item{Frecuencia alelica locus B/b}{Un vector con las frecuencias alelicas para el locus B/b (p_B y q_B).}
\item{Frecuencia alelica Haplotipo}{Un vector con las frecuencias de los haplotipos AB, Ab, aB y ab.}
}
}
\description{
Esta funcion calcula las frecuencias genotipicas y alelicas en un marcador de dos loci (dilocus),
dada una tabla con los genotipos y sus frecuencias observadas. Ademas, calcula las frecuencias de los haplotipos combinados
para los loci involucrados.
}
\examples{

# Datos de entrada
dilocus_table <- data.frame(
 Pop = c("Cur1", "ANT2"),
 AABB = c(0, 1),
 AABb = c(0, 2),
 AAbb = c(0, 0),
 AaBB = c(38, 30),
 AaBb = c(0, 0),
 Aabb = c(0, 0),
 aaBB = c(2, 3),
 aaBb = c(0, 0),
 aabb = c(0, 0))

 # Calcular las frecuencias dilocus para cada población
 Df2 <- Freq_All_Diloc(dilocus_table)

 # Mostrar los resultados
 print(Df2)
}
