% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenDiLoc.R
\name{GenDiLoc}
\alias{GenDiLoc}
\title{Calcular frecuencia genotípica de un marcador dilocus}
\usage{
GenDiLoc(Data)
}
\arguments{
\item{Data}{Un data frame que contiene los datos geneticos. Debe incluir al menos
las columnas \code{Pop} (identificador de la población) y las columnas de genotipos:
\code{AA}, \code{Aa}, \code{aa}, \code{BB}, \code{Bb}, \code{bb}, que indican la presencia de cada combinación alelica.}
}
\value{
Una matriz en la que cada fila corresponde a una población y cada columna
representa un genotipo posible (\code{AABB}, \code{AABb}, etc.). Las celdas contienen las frecuencias
absolutas (conteos) de cada genotipo en cada población.
}
\description{
Esta función calcula las frecuencias genotípicas para un marcador genético dilocus
en una población, basado en los genotipos observados en los individuos.
}
\examples{
# Not Run sin utilizar la funcion Import_di_loc
df <- data.frame(
  ID = 1:12,
  Pop = rep("Pop1", 12),
  AA = c(0, 0, 0, 0, "X", 0, 0, 0, 0, 0, 0, "X"),
  Aa = c("X", 0, "X", "X", 0, "X", "X", "X", "X", "X", "X", 0),
  aa = c(0, "X", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
  BB = c("X", 0, 0, 0, 0, 0, "X", 0, 0, 0, 0, 0),
  Bb = c(0, "X", "X", "X", 0, "X", 0, 0, "X", "X", "X", "X"),
  bb = c(0, 0, 0, 0, "X", 0, 0, "X", 0, 0, 0, 0))

# Not Run
GenDiLoc(df)

}
