% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hap_DilocusfreqforYear.R
\name{Hap_DilocusfreqforYear}
\alias{Hap_DilocusfreqforYear}
\title{Funcion para calcular la frecuencias haplotipica con mutaciones Dilocus}
\usage{
Hap_DilocusfreqforYear(data, pais)
}
\arguments{
\item{data}{Datos originales se puede utilizar V1016I.F1535C o V1016G.F1535C}

\item{pais}{Pais del cual deseamos extraer los datos}
}
\value{
Tabla con las frecuencias haplotipicas del pais por año
}
\description{
Funcion para calcular la frecuencias haplotipica con mutaciones Dilocus
}
\examples{
Freq_hap_mexico <- Hap_DilocusfreqforYear(V1016I.F1534C, "Mexico")

}
