% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/HW.r
\name{HW}
\alias{HW}
\title{Finds proportion of genotypes pp, pq and qq, given P (i.e. sqrt(pp))}
\usage{
HW(P, mat)
}
\arguments{
\item{P}{todo}

\item{mat}{todo}
}
\value{
matrix
}
\description{
Takes P and matrix to print results to, returning matrix
}

