% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWQL1.R
\name{HWQL1}
\alias{HWQL1}
\title{Calcular el Test de Hardy-Weinberg}
\usage{
HWQL1(genotype_counts)
}
\arguments{
\item{genotype_counts}{Un vector numerico que contiene los conteos de los genotipos observados.
Debe tener tres elementos, correspondientes a los genotipos \code{AA}, \code{Aa} y \code{aa} en este orden.}
}
\value{
Un lista con los siguientes elementos:
\item{Total_Individuals}{Número total de individuos observados en la muestra.}
\item{Genotypic_Frequencies}{Un data frame con las frecuencias genotípicas observadas.}
\item{Allelic_Frequencies}{Un data frame con las frecuencias alélicas observadas.}
\item{Expected_Frequencies}{Un data frame con las frecuencias esperadas según Hardy-Weinberg para los genotipos \code{AA}, \code{Aa} y \code{aa}, y sus correspondientes conteos esperados.}
\item{Observed_Counts}{Un data frame con los conteos observados de los genotipos \code{AA}, \code{Aa} y \code{aa}.}
\item{Chi_Squared}{Valor del estadístico chi-cuadrado calculado para la prueba de ajuste Hardy-Weinberg.}
\item{P_Value}{P-valor asociado al estadístico chi-cuadrado para determinar si la diferencia entre las frecuencias observadas y esperadas es significativa.}
}
\description{
Esta función realiza un test de Hardy-Weinberg utilizando los conteos de genotipos observados.
Calcula las frecuencias genotípicas y alélicas observadas y esperadas según Hardy-Weinberg,
luego evalúa si la población está en equilibrio usando el estadístico de chi-cuadrado y su p-valor.
}
\examples{
# Ejemplo con un conjunto de datos de conteos genotipicos para una sola poblacion
df1 <- data.frame(
 Pop = c("Pop1"),
 aa = c(8),
 Aa = c(43),
 AA = c(25))

HWQL1(df1)


# Ejemplo de uso con un conjunto de datos de conteos genotípicos para múltiples poblaciones

df2 <- data.frame(
 Pop = c("Pop 1", "Pop 2", "Pop 3"),
 AA = c(4, 0, 1),
 Aa = c(33, 39, 22),
 aa = c(3, 2, 7))

 HWQL1(df2)

}
