% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HWQL2.R
\name{HWQL2}
\alias{HWQL2}
\title{Realiza un Test de Equilibrio de Hardy-Weinberg para múltiples poblaciones con marcadores dihíbridos (dos loci)}
\usage{
HWQL2(genotype_counts, gl = 1)
}
\arguments{
\item{genotype_counts}{Un data frame que contiene los conteos de genotipos observados
para múltiples poblaciones. Debe incluir las siguientes columnas:
\itemize{
\item \code{Pop}: Identificador de la población.
\item \code{AABB}, \code{AABb}, \code{AAbb}, \code{AaBB}, \code{AaBb}, \code{Aabb},
\code{aaBB}, \code{aaBb}, \code{aabb}: Conteos de los genotipos para los dos loci.
}}

\item{gl}{Grados de libertad para la prueba de chi-cuadrado. Por defecto es 1.}
}
\value{
Un data frame con los siguientes resultados para cada población:
\itemize{
\item \code{Population}: Identificador de la población.
\item \code{n}: Número total de individuos en la población.
\item \code{AB}, \code{Ab}, \code{aB}, \code{ab}: Frecuencias haplotípicas estimadas
para los haplotipos AB, Ab, aB y ab.
\item \code{X2 (df)}: Valor del estadístico chi-cuadrado y grados de libertad.
\item \code{P_Value}: P-valor asociado al estadístico chi-cuadrado.
}
}
\description{
Esta función calcula las frecuencias genotípicas y haplotípicas observadas y esperadas
según el equilibrio de Hardy-Weinberg para múltiples poblaciones con marcadores dihíbridos
(dos loci). Utiliza el método de iteración de Expectation-Maximization (EM) para estimar
las frecuencias haplotípicas y realiza una prueba de chi-cuadrado para evaluar el equilibrio
de Hardy-Weinberg. \strong{Esta función está diseñada específicamente para marcadores dihíbridos},
es decir, dos loci con dos alelos cada uno (A/a y B/b), lo que genera genotipos como AABB,
AABb, AAbb, AaBB, AaBb, Aabb, aaBB, aaBb y aabb.
}
\details{
La función utiliza la función \code{Diloc_Freq_Hd} para estimar las frecuencias
haplotípicas mediante el método de iteración EM. Luego, calcula las frecuencias
genotípicas esperadas bajo el equilibrio de Hardy-Weinberg y realiza una prueba de
chi-cuadrado para comparar las frecuencias observadas y esperadas.

Si una población tiene menos de dos genotipos con conteos mayores que cero, se omite
y se muestra una advertencia.
}
\examples{
# Cargar datos de ejemplo
data <- Import_di_loc(Loc2Pops3)

# Realizar el test de Hardy-Weinberg con 3 grados de libertad
resultados <- HWQL2(data$DicGtps, gl = 3)

# Mostrar los resultados
print(resultados)

}
