% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impor_Single_Locus.R
\name{Impor_Single_Locus}
\alias{Impor_Single_Locus}
\title{Importa datos de un solo locus}
\usage{
Impor_Single_Locus(df)
}
\arguments{
\item{df}{Un data.frame llamado "Matriz de Genotipos" que representa la presencia de los genotipos (aa, Aa, AA) para un conjunto de individuos identificados por su ID.}
}
\value{
Un vector numerico con el conteo de individuos para cada genotipo en el orden: AA, Aa, aa.
}
\description{
Esta funcion verifica la consistencia de una matriz de genotipos y cuenta el numero de individuos con cada genotipo (aa, Aa, AA) en un conjunto de datos.
}
\examples{


genotipos <- data.frame(
  ID = 1:12,
  aa = c(0, 0, 0, 0, "X", 0, 0, 0, 0, 0, 0, "X"),
  Aa = c("X", 0, "X", "X", 0, "X", "X", "X", "X", "X", "X", 0),
  AA = c(0, "X", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))

Impor_Single_Locus(genotipos)

}
