% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_di_loc.R
\name{Import_di_loc}
\alias{Import_di_loc}
\title{Import two Locus Data Mutation KDR AA/Aa/aa and BB/Bb/bb}
\usage{
Import_di_loc(df)
}
\arguments{
\item{df}{A data.frame called "Genotype Matrix" that represents the presence of genotypes (AA, Aa, aa, BB, Bb, bb) for a set of individuals identified by their ID and population.}
}
\value{
A data.frame with the frequency of each genotype (AA, Aa, aa, BB, Bb, bb) per population.
}
\description{
This function checks the consistency of a genotype matrix and counts the number of individuals with each genotype (AA, Aa, aa, BB, Bb, bb) in a dataset. It is specifically designed for working with KDR mutation genotypes and an additional locus for BB/Bb/bb genotypes, often used in resistance studies.
}
\details{
The function performs several checks to ensure data integrity:
\itemize{
\item Validates that the input is a \code{data.frame}
\item Ensures the presence of the required columns: \code{ID}, \code{Pop}, \code{AA}, \code{Aa}, \code{aa}, \code{BB}, \code{Bb}, and \code{bb}.
\item Checks for the absence of \code{NA} values in the required columns.
\item Ensures that each genotype column only contains the values: \code{X} (indicating the individual has the genotype) and \code{0} (indicating absence of the genotype).
\item Ensures that each individual (row) has exactly one \code{X}, meaning an individual can only present one genotype.
\item Validates that the total count of \code{X} matches the total number of individuals (rows) in the data frame.
}

The function counts the number of \code{X} for each genotype in each population and returns a frequency matrix of genotypes by population.
}
\examples{

# Example 1: Provides a matrix indicating the genotype for 2 loci for 12 individuals
# from 1 population

df <- data.frame(
  ID = 1:12,
  Pop = rep("Pop1", 12),
  AA = c(0, 0, 0, 0, "X", 0, 0, 0, 0, 0, 0, "X"),
  Aa = c("X", 0, "X", "X", 0, "X", "X", "X", "X", "X", "X", 0),
  aa = c(0, "X", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
  BB = c("X", 0, 0, 0, 0, 0, "X", 0, 0, 0, 0, 0),
  Bb = c(0, "X", "X", "X", 0, "X", 0, 0, "X", "X", "X", "X"),
  bb = c(0, 0, 0, 0, "X", 0, 0, "X", 0, 0, 0, 0))

Import_di_loc(df)


# Example 2: Provides a matrix indicating the genotype for 106 individuals in 3 populations

Import_di_loc(Loc2Pops3)

}
