% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_single_locus.R
\name{Import_single_loc}
\alias{Import_single_loc}
\title{Import Single Locus Data Mutation KDR}
\usage{
Import_single_loc(df)
}
\arguments{
\item{df}{A data.frame called "Genotype Matrix" that represents the presence of genotypes (aa, Aa, AA) for a set of individuals identified by their ID.}
}
\value{
A numeric vector with the count of individuals for each genotype in the order: AA, Aa, aa.
}
\description{
This function checks the consistency of a genotype matrix and counts the number of individuals with each genotype (aa, Aa, AA) in a dataset. It is specifically designed for working with KDR mutation genotypes that confer resistance to pyrethroids.
}
\details{
The function performs several checks to ensure data integrity:
\itemize{
\item Validates that the input is a \code{data.frame}.
\item Ensures the presence of the required columns: \code{ID}, \code{aa}, \code{Aa}, and \code{AA}.
\item Checks for the absence of \code{NA} values in the required columns.
\item Ensures that each genotype column only contains the values: \code{X} (indicating the individual has the genotype) and \code{0} (indicating absence of the genotype).
\item Ensures that each individual (row) has exactly one \code{X}, meaning an individual can only present one genotype.
\item Validates that the total count of \code{X} matches the total number of individuals (rows) in the data frame.
}
}
\examples{

# Example 1: Provides a matrix indicating the genotype for 12 individuals for 1 population

df <- data.frame(
  ID = 1:12,
  Pop = rep("Pop1", 12),
  AA = c(0, 0, 0, 0, "X", 0, 0, 0, 0, 0, 0, "X"),
  Aa = c("X", 0, "X", "X", 0, "X", "X", "X", "X", "X", "X", 0),
  aa = c(0, "X", 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))


Import_single_loc(df)


# Example 2: Provides a matrix indicating the genotype for 106 individuals in 3 populations

Import_single_loc(Loc1Pops3)

}
