% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loc1Pops3.R
\docType{data}
\name{Loc1Pops3}
\alias{Loc1Pops3}
\title{Database de conteo de genotipos aa, Aa y AA de 111 individuos agrupados en 3 poblaciones.}
\format{
\subsection{\code{Loc1Pops3}}{

A data frame with 106 rows and 5 columns:
\describe{
\item{ID}{Identificador individual}
\item{Pop}{Población a la que pertenece cada individuo}
\item{aa}{Frecuencia homocigoto recesivo}
\item{Aa}{Frecuencia heterocigoto}
\item{AA}{Frecuencia homocigoto dominante}
}
}
}
\source{
<Datos resultados Artículo KDR Paraná, Brasil>
}
\usage{
Loc1Pops3
}
\description{
Database de conteo de genotipos aa, Aa y AA de 111 individuos agrupados en 3 poblaciones.
}
\keyword{datasets}
