% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Loc2Pops3.R
\docType{data}
\name{Loc2Pops3}
\alias{Loc2Pops3}
\title{Database de conteo de genotipos AA, Aa, aa, BB, Bb y bb de 111 individuos agrupados en 3 poblaciones.}
\format{
\subsection{\code{Loc2Pops3}}{

A data frame with 111 rows and 5 columns:
\describe{
\item{ID}{Identificador individual}
\item{Pop}{Población a la que pertenece cada individuo}
\item{aa}{Frecuencia homocigoto dominante for loci 1}
\item{Aa}{Frecuencia heterocigoto for loci 1}
\item{AA}{Frecuencia homocigoto recesivo for loci 1}
\item{BB}{Frecuencia homocigoto dominante for loci 1}
\item{Ba}{Frecuencia heterocigoto for loci 1}
\item{bb}{Frecuencia homocigoto recesivo for loci 1}
}
}
}
\source{
<Datos resultados Artículo KDR Paraná, Brasil>
}
\usage{
Loc2Pops3
}
\description{
Database de conteo de genotipos AA, Aa, aa, BB, Bb y bb de 111 individuos agrupados en 3 poblaciones.
}
\keyword{datasets}
