% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.genotypemat.r
\name{make.genotypemat}
\alias{make.genotypemat}
\title{Uses HW function to make matrix of genotype values per matrix}
\usage{
make.genotypemat(P_1, P_2)
}
\arguments{
\item{P_1}{todo}

\item{P_2}{todo}
}
\value{
matrix
}
\description{
Takes frequencies of resistant allele at each locus - user inputted
}

