% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/make.matrix.r
\name{make.matrix}
\alias{make.matrix}
\title{Makes data.frame into a matrix with parameters as rownames}
\usage{
make.matrix(mat, rnames)
}
\arguments{
\item{mat}{}

\item{rnames}{}
}
\value{
? dat.m
}
\description{
takes data frame as dat, and column in data frame containing parameters as rnames.
this produces NA generation errors due to coercing location column, this column of NA's is however removed.
}

