% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotallele.freq.r
\name{plotallele.freq}
\alias{plotallele.freq}
\title{plot allele frequencies over time}
\usage{
plotallele.freq(mat)
}
\arguments{
\item{mat}{results matrix}
}
\value{
list of plot components
}
\description{
plot of R and S allele frequencies over time
Prints male frequency of R allele at locus 1 (blue) and locus 2 (green)
and same in female at locus 1 (red) and locus 2 (orange)
}

