% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAlleleFrequency.R
\name{PlotAlleleFrequency}
\alias{PlotAlleleFrequency}
\title{Plot Allele Frequencies by Population}
\usage{
PlotAlleleFrequency(data, Loc = 2)
}
\arguments{
\item{data}{A data frame containing the population, allele, and frequency data.}

\item{Loc}{An integer (1 or 2) indicating the type of input data structure.
\itemize{
\item{1: Data structure with alleles "A" and "a".}
\item{2: Data structure with haplotypes "Ab", "AB", "aB", and "ab".}
}}
}
\value{
A ggplot object representing the bar plot of allele frequencies by population.
}
\description{
This function creates a bar plot of allele frequencies by population. It can handle two types of input data structures,
specified by the Loc argument.
}
\examples{

# Example 1 for Loc = 1
data1 <- data.frame(
  Pop = c("Pop1", "Pop2", "Pop3", "Pop4", "Pop5", "Pop6"),
  Freq_A = c(0.6, 0.7, 0.5, 0.8, 0.4, 0.3),
  Freq_a = c(0.4, 0.3, 0.5, 0.2, 0.6, 0.7)
)

dataf1 <- RestructureAlleleFreq(data1, Loc = 1)
PlotAlleleFrequency(dataf1, Loc = 1)

# Example 2 for Loc = 2
data2 <- data.frame(
  Pop = c("Pop1", "Pop2", "Pop3", "Pop4", "Pop5", "Pop6"),
  AB = c(0.2, 0.3, 0.1, 0.25, 0.4, 0.35),
  Ab = c(0.3, 0.4, 0.2, 0.15, 0.25, 0.3),
  aB = c(0.4, 0.2, 0.3, 0.45, 0.2, 0.25),
  ab = c(0.1, 0.1, 0.4, 0.15, 0.15, 0.1)
)

dataf2 <- RestructureAlleleFreq(data2, Loc = 2)
PlotAlleleFrequency(dataf2, Loc = 2)
}
