% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotallele.freq.andy.r
\name{plotalleleLinesOneScenario}
\alias{plotalleleLinesOneScenario}
\title{helper function to plot allele frequency lines for one scenario}
\usage{
plotalleleLinesOneScenario(mat)
}
\arguments{
\item{mat}{results matrix}
}
\value{
nothing
}
\description{
helper function to plot allele frequency lines for one scenario
}

