% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotcurtis_f1.r
\name{plotcurtis_f1}
\alias{plotcurtis_f1}
\title{plot Resistance Allele frequencies over time}
\usage{
plotcurtis_f1(nrelaxmat, relaxmat, gencol, r1col)
}
\arguments{
\item{nrelaxmat}{results matrix without relaxation of selection}

\item{relaxmat}{results matrix with relaxation of selection}

\item{gencol}{column number containing generation}

\item{r1col}{column number containing frequency of allele at R1}
}
\value{
list of plot components
}
\description{
Recreates figure 1 of Curtis(1985)
Plot of R at each locus (R1=R2, as in Curtis fig 1)
Prints total frequency of R allele at locus 1 (red) and locus 2 (blue)
Points plotted in purple indicate generations where selection is relaxed
}

