% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotcurtis_f2.r
\name{plotcurtis_f2}
\alias{plotcurtis_f2}
\title{plot Resistance Allele frequencies over time}
\usage{
plotcurtis_f2(combmat, bmat, amat, gencol, r1col, r2col)
}
\arguments{
\item{combmat}{todo}

\item{bmat}{todo}

\item{amat}{todo}

\item{gencol}{column number containing generation}

\item{r1col}{column number containing frequency of allele at R1}

\item{r2col}{column number containing frequency of allele at R2}
}
\value{
list of plot components
}
\description{
Recreates figure 2 of Curtis(1985)
Plot of R at each locus
Prints total frequency of R allele at locus 1 (red) and locus 2 (blue)
}

