% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotcurtis_ld.r
\name{plotcurtis_ld}
\alias{plotcurtis_ld}
\title{plot linkage disequilibrium over time}
\usage{
plotcurtis_ld(nrelaxmat, relaxmat, gencol, ldcol)
}
\arguments{
\item{nrelaxmat}{results matrix without relaxation of selection}

\item{relaxmat}{results matrix with relaxation of selection}

\item{gencol}{column number containing generation (1)}

\item{ldcol}{column number containing linkage disequilibrium (4)}
}
\value{
list of plot components
}
\description{
Points plotted in purple indicate generations where selection is relaxed
}

