% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHapDilocusFreq.R
\name{PlotHapDilocusFreq}
\alias{PlotHapDilocusFreq}
\title{Grafica la frecuencia haplotípica de mutaciones dilocus por año.}
\usage{
PlotHapDilocusFreq(data, pais)
}
\arguments{
\item{data}{Datos de la frecuencias haplotípica de mutaciones dilocus a graficar.}

\item{pais}{País o localidad que se desea graficar.}
}
\value{
Gráfico con la frecuencia haplotípica de mutaciones dilocus por año.
}
\description{
Grafica la frecuencia haplotípica de mutaciones dilocus por año.
}
\examples{
PlotHapDilocusFreq(V1016I.F1534C, "Mexico")

}
