% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plothaplotype.r
\name{plothaplotype}
\alias{plothaplotype}
\title{plot Haplotype}
\usage{
plothaplotype(mat)
}
\arguments{
\item{mat}{results matrix}
}
\value{
list of plot components
}
\description{
Plots numbers of homozygote S, heterozygote and homozygote R at each generation (from genotype matrix)
locus 1: SS in pink, RS in orange, RR in red
locus 2: SS in cyan, RS in dark blue, RR in green
}

