% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotlinkage.r
\name{plotlinkage}
\alias{plotlinkage}
\title{plot Linkage Disequilibrium over generations}
\usage{
plotlinkage(mat)
}
\arguments{
\item{mat}{results matrix}
}
\value{
list of plot components
}
\description{
plot Linkage Disequilibrium over generations
}

