% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resistSimple.r
\name{resistSimple}
\alias{resistSimple}
\title{to run one simple scenario without an input file, experimental}
\usage{
resistSimple(calibration = 1012, produce.plots = TRUE, savePlots = FALSE,
  save.fitvals = 0, ...)
}
\arguments{
\item{calibration}{one of a limited set of integers effecting how scenarios are run}

\item{produce.plots}{whether to produce plots}

\item{savePlots}{whether to save plots to hardcoded filenames}

\item{save.fitvals}{whether to save fitness scores to csv 0=no, 1=yes}

\item{...}{extra arguments passed to setInputOneScenario()}
}
\value{
nothing yet
}
\description{
May restructure later.
}

