% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runModel.r
\name{runModel}
\alias{runModel}
\title{run the model scenarios specified in the input object}
\usage{
runModel(input, calibration, produce.plots = TRUE, savePlots = FALSE)
}
\arguments{
\item{input}{a matrix with parameters in rows and scenarios in columns}

\item{calibration}{one of a limited set of integers effecting how scenarios are run}

\item{produce.plots}{whether to produce plots}

\item{savePlots}{whether to save plots to hardcoded filenames}
}
\value{
a list of 3 lists of one or more scenarios: results, genotype and fitness. e.g. listOut$results[1] gives a results matrix for the first scenario
}
\description{
run model scenarios and put the results for each in a results object
}

