% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runUI1.R
\name{runUI1}
\alias{runUI1}
\title{run a shiny User Interface to a simple version of the resistance model used by curtis}
\usage{
runUI1()
}
\value{
nothing
}
\description{
displays a user interface in the browser, allowing users to modify inputs and see how the frequency of resistance alleles responds
}

