% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sensiAn1.r
\name{sensiAn1}
\alias{sensiAn1}
\title{sensitivity analysis first attempt}
\usage{
sensiAn1(nScenarios = 10, ...)
}
\arguments{
\item{nScenarios}{number of scenarios}

\item{...}{parameter ranges to pass to \code{\link{setInputSensiScenarios}} in form c(min,max)}
}
\value{
todo
}
\description{
sensitivity analysis first attempt
}
\examples{
sensiAn1( P_1=c(0,0.5) )
inputAndResults <- sensiAn1( h.RS1_A0=c(0.1,1))
colnames(inputAndResults)
inputAndResults <- sensiAn1( h.RS1_A0=c(0.1,1), h.RS2_0B=c(0.1,1) )
inputAndResults <- sensiAn1(500, h.RS1_A0=c(0.1,1), h.RS2_0B=c(0.1,1), s.RR1_A0=c(0.2,1), s.RR2_0B=c(0.2,1))
}

