% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setInputOneScenario.r
\name{setInputOneScenario}
\alias{setInputOneScenario}
\title{set inputs for one scenario}
\usage{
setInputOneScenario(calibration = 100, max_gen = 100, coll.fitvals = 1,
  save.fitvals = 0, P_1 = 0.001, P_2 = 0.001, recomb_rate = 0.5,
  a.m_00 = 0.1, a.m_a0 = 0, a.m_A0 = 0, a.m_0b = 0, a.m_0B = 0,
  a.m_ab = 0, a.m_AB = 0.9, a.m_Ab = 0, a.m_aB = 0, a.f_00 = 0.1,
  a.f_a0 = 0, a.f_A0 = 0, a.f_0b = 0, a.f_0B = 0, a.f_ab = 0,
  a.f_AB = 0.9, a.f_Ab = 0, a.f_aB = 0, phi.SS1_a0 = 0,
  phi.SS1_A0 = 1, phi.SS2_0b = 0, phi.SS2_0B = 1, W.SS1_00 = 1,
  W.SS2_00 = 1, h.RS1_00 = 0, h.RS1_a0 = 0, h.RS1_A0 = 1,
  h.RS2_00 = 0, h.RS2_0b = 0, h.RS2_0B = 1, s.RR1_a0 = 0,
  s.RR1_A0 = 1, s.RR2_0b = 0, s.RR2_0B = 1, z.RR1_00 = 0,
  z.RR2_00 = 0, niche_00 = 1, niche_a0 = 0, niche_A0 = 0,
  niche_0b = 0, niche_0B = 0, niche_ab = 0, niche_AB = 1,
  niche_Ab = 0, niche_aB = 0)
}
\arguments{
\item{calibration}{one of a limited set of integers effecting how scenarios are run}

\item{max_gen}{number of generations}

\item{coll.fitvals}{save fitness values in a matrix 0/1}

\item{save.fitvals}{save fitness values to an external .csv 0/1}

\item{P_1}{locus 1 frequency of resistance allele}

\item{P_2}{locus 2 frequency of resistance allele}

\item{recomb_rate}{recombination rate}

\item{a.m_00}{insecticide exposure male no1 no2}

\item{a.m_a0}{insecticide exposure male lo1 no2}

\item{a.m_A0}{insecticide exposure male hi1 no2}

\item{a.m_0b}{insecticide exposure male no1 lo2}

\item{a.m_0B}{insecticide exposure male no1 hi2}

\item{a.m_ab}{insecticide exposure male lo1 lo2}

\item{a.m_AB}{insecticide exposure male hi1 hi2}

\item{a.m_Ab}{insecticide exposure male hi1 lo2}

\item{a.m_aB}{insecticide exposure male lo1 hi2}

\item{a.f_00}{insecticide exposure female no1 no2}

\item{a.f_a0}{insecticide exposure female lo1 no2}

\item{a.f_A0}{insecticide exposure female hi1 no2}

\item{a.f_0b}{insecticide exposure female no1 lo2}

\item{a.f_0B}{insecticide exposure female no1 hi2}

\item{a.f_ab}{insecticide exposure female lo1 lo2}

\item{a.f_AB}{insecticide exposure female hi1 hi2}

\item{a.f_Ab}{insecticide exposure female hi1 lo2}

\item{a.f_aB}{insecticide exposure female lo1 hi2}

\item{phi.SS1_a0}{Baseline fitness of SS1 in a0}

\item{phi.SS1_A0}{Baseline fitness of SS1 in A0}

\item{phi.SS2_0b}{Baseline fitness of SS2 in 0b}

\item{phi.SS2_0B}{Baseline fitness of SS2 in 0B}

\item{W.SS1_00}{Fitness of SS1 in 00 (no insecticide)}

\item{W.SS2_00}{Fitness of SS2 in 00 (no insecticide)}

\item{h.RS1_00}{Dominance coefficient locus1 in 00}

\item{h.RS1_a0}{Dominance coefficient locus1 in a0}

\item{h.RS1_A0}{Dominance coefficient locus1 in A0}

\item{h.RS2_00}{Dominance coefficient locus2 in 00}

\item{h.RS2_0b}{Dominance coefficient locus2 in 0b}

\item{h.RS2_0B}{Dominance coefficient locus2 in 0B}

\item{s.RR1_a0}{Selection coefficient locus1 in a}

\item{s.RR1_A0}{Selection coefficient locus1 in A}

\item{s.RR2_0b}{Selection coefficient locus2 in b}

\item{s.RR2_0B}{Selection coefficient locus2 in B}

\item{z.RR1_00}{fitness cost of resistance allele 1 in insecticide free environment}

\item{z.RR2_00}{fitness cost of resistance allele 2 in insecticide free environment}

\item{niche_00}{insecticide niche toggle no1 no2 0=off 1=on}

\item{niche_a0}{insecticide niche toggle no1 no2 0=off 1=on}

\item{niche_A0}{insecticide niche toggle hi1 no2 0=off 1=on}

\item{niche_0b}{insecticide niche toggle no1 lo2 0=off 1=on}

\item{niche_0B}{insecticide niche toggle no1 hi2 0=off 1=on}

\item{niche_ab}{insecticide niche toggle lo1 lo2 0=off 1=on}

\item{niche_AB}{insecticide niche toggle hi1 hi2 0=off 1=on}

\item{niche_Ab}{insecticide niche toggle hi1 lo2 0=off 1=on}

\item{niche_aB}{insecticide niche toggle lo1 hi2 0=off 1=on}
}
\value{
named vector
}
\description{
fills one column of input matrix
user can specify as many or few parameters as they wish, any not specified will be set to default value.
}
\details{
BEWARE function programmer not to change order of the function arguments as this determines the names of the output
}

