% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/setInputSensiScenarios.r
\name{setInputSensiScenarios}
\alias{setInputSensiScenarios}
\title{set inputs for a sensitivity analysis of multiple scenarios}
\usage{
setInputSensiScenarios(nScenarios = 10, ...)
}
\arguments{
\item{nScenarios}{number of scenarios to generate}
}
\value{
a matrix of input scenarios
}
\description{
creates a multi-column input matrix which can be passed to runModel()
You need to pass a range for which parameters you want, as c(min,max).
Scenarios will be created by selecting from a uniform distribution specified by this range.
Any parameters not specified will be set to their default values.
TODO: I might be able to allow parameters to be set to non default values from here too.
}
\examples{
setInputSensiScenarios( nScenarios=2, h.RS1_A0=c(0,1) )
setInputSensiScenarios( nScenarios=3, P_1=c(0,0.5), P_2=c(0,0.5) )
}

