% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_loc_freq.R
\name{Single_loc_freq}
\alias{Single_loc_freq}
\title{Calculate Genotypic and Allelic Frequencies}
\usage{
Single_loc_freq(freq_matrix)
}
\arguments{
\item{freq_matrix}{A data frame with the following columns:
\describe{
\item{Pop}{A unique identifier for each population (e.g., "Pop1", "Pop2").}
\item{aa}{The count of individuals with the genotype aa.}
\item{Aa}{The count of individuals with the genotype Aa.}
\item{AA}{The count of individuals with the genotype AA.}
}}
}
\value{
A data frame with the following columns:
\describe{
\item{Population}{The identifier for each population.}
\item{n}{The total number of individuals considered in the calculations.}
\item{aa}{The genotypic frequency of individuals with the aa genotype.}
\item{Aa}{The genotypic frequency of individuals with the Aa genotype.}
\item{AA}{The genotypic frequency of individuals with the AA genotype.}
\item{Freq_A}{The allelic frequency of allele A.}
\item{Freq_a}{The allelic frequency of allele a.}
}
}
\description{
This function calculates the genotypic and allelic frequencies from a matrix containing genotype counts (AA, Aa, aa) for each population.
}
\examples{
# Example: Genotypic frequency matrix for a single population
freq_matrix <- data.frame(
  Pop = c("Pop1"),
  aa = c(2),
  Aa = c(9),
  AA = c(1))

# Calculate frequencies
Single_loc_freq(freq_matrix)

# Example: Genotypic frequency matrix for multiple populations
df2 <- Import_single_loc(Loc1Pops3)
Single_loc_freq(df2)

}
