% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_locus_frequencies.R
\name{Single_locus_frequencies}
\alias{Single_locus_frequencies}
\title{Calcular frecuencias genotipicas y alelicas sin locus}
\usage{
Single_locus_frequencies(genotype_counts)
}
\arguments{
\item{genotype_counts}{Un vector numerico de longitud 3 que representa los conteos de individuos con los genotipos: AA, Aa y aa, en ese orden.}
}
\value{
Una lista con tres elementos:
\describe{
\item{genotypic_frequencies}{Un data.frame con las frecuencias genotipicas de AA, Aa, y aa.}
\item{allelic_frequencies}{Un data.frame con las frecuencias alelicas de A y a.}
\item{total_individuals}{El numero total de individuos considerados en los calculos.}
}
}
\description{
Esta funcion calcula las frecuencias genotipicas y alelicas a partir de un vector que contiene los conteos de individuos para cada genotipo (AA, Aa, aa).
}
\examples{

genotype_counts <- c(AA=50, Aa=30, aa=20)  # 50 individuos con genotipo AA, 30 con Aa, y 20 con aa

result <- Single_locus_frequencies(genotype_counts)

print(result$genotypic_frequencies) # Frecuencias genotipicas
print(result$allelic_frequencies)  # Frecuencias alelicas
print(result$total_individuals)    # Total de individuos

}
