% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/singlealleleFrequency.r
\name{singlealleleFrequency}
\alias{singlealleleFrequency}
\title{Return allele frequencies}
\usage{
singlealleleFrequency(locus, max_gen, results.list, input)
}
\arguments{
\item{locus}{locus of interest (i.e. 1 or 2)}

\item{max_gen}{number of generations
(should be the maximum number used across all runs, i.e. the longest any run is set for)}

\item{results.list}{list containing the results matrices generated for each scenario run}

\item{input}{todo}
}
\value{
matrix with rows for each generation, and columns for each scenario
}
\description{
extract just the R allele frequencies at a specified locus, for all scenarios run
}

