% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timetoFifty.r
\name{timetoFifty}
\alias{timetoFifty}
\title{Find number of generations to 0.5}
\usage{
timetoFifty(locus, max_gen, results.list, input)
}
\arguments{
\item{locus}{locus of interest (i.e. 1 or 2)}

\item{max_gen}{number of generations
(should be the maximum number used across all runs, i.e. the longest any run is set for)}

\item{results.list}{list containing the results matrices generated for each scenario run}

\item{input}{todo}
}
\value{
matrix with a column for each scenario, giving the generation number at which frequency of R is first >= 0.5
}
\description{
Takes the matrix produced by single allele frequency function
and finds the number of generations it takes for the R allele frequency to reach 0.5
Returns NA if freq does not reach 0.5
}

