% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V1016I.F1534C.R
\docType{data}
\name{V1016I.F1534C}
\alias{V1016I.F1534C}
\title{Database of the global and temporal compilation of the distribution and frequency by countries and locations of the dilocus mutation V1016I + F1534C.}
\format{
\subsection{\code{V1016I.F1534C}}{

A data frame with 582 rows and 26 columns:
\describe{
\item{ID}{Identificardor individual}
\item{Year}{Año del estudio}
\item{Title}{Titulo del Articulo donde se obtuvo la informaciòn}
}
}
}
\source{
<Datos resultados de una revisiòn sistematizada hasta el 2023>
}
\usage{
V1016I.F1534C
}
\description{
Database of the global and temporal compilation of the distribution and frequency by countries and locations of the dilocus mutation V1016I + F1534C.
}
\keyword{datasets}
