% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createArray.r
\name{createArray}
\alias{createArray}
\title{deprecated, replaced with createArray2() create array for representing genotypes and niches}
\usage{
createArray(sex = NULL, loci = NULL, locus1 = NULL, locus2 = NULL,
  niche1 = NULL, niche2 = NULL, exposure = NULL, fillValue = 0)
}
\arguments{
\item{sex}{vector of dimension names e.g. c('f','m')}

\item{loci}{vector of dimension names e.g. c('SS1','RS1','RR1','SS2','RS2','RR2')}

\item{locus1}{vector of dimension names e.g. c('SS1','RS1','RR1')}

\item{locus2}{vector of dimension names e.g. c('SS2','RS2','RR2')}

\item{niche1}{vector of dimension names e.g. c('0','a','A')}

\item{niche2}{vector of dimension names e.g. c('0','a','A')}

\item{exposure}{vector of dimension names e.g. c('no','lo','hi')}

\item{fillValue}{optional what to fill array with, defaults to 0}
}
\value{
matrix
}
\description{
creates array with the passed dimensions
}
\examples{
createArray( sex = c('f','m') )
#Wloci
createArray( loci=c('SS1','RS1','RR1','SS2','RS2','RR2'), niche1=c('0','a','A'), niche2=c('0','b','B') )
createArray( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2') )
createArray( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2'), niche1 = c('0','a','A') )
createArray( sex = c('f','m'), locus1 = c('SS1','RS1','RR1'), locus2 = c('SS2','RS2','RR2'), niche1 = c('0','a','A'), niche2 = c('0','a','A') )
}

