% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diloc_Freq_Hd.R
\name{Diloc_Freq_Hd}
\alias{Diloc_Freq_Hd}
\title{Calculate Allele Frequencies for a Biallelic Marker Across Multiple Populations}
\usage{
Diloc_Freq_Hd(data, max_iterations = 10, tol = 1e-06, dat = 1)
}
\arguments{
\item{data}{A data frame where each row corresponds to a different population. The first column should contain the population names
(\code{Pop}), and the remaining columns should represent the genotype counts for the following genotypes: \code{AABB}, \code{AABb}, \code{AAbb}, \code{AaBB},
\code{AaBb}, \code{Aabb}, \code{aaBB}, \code{aaBb}, \code{aabb}.}

\item{max_iterations}{The maximum number of iterations to perform in the estimation process. Default is 15.}

\item{tol}{The tolerance level for convergence. If the change in the correction factors (\code{CorAB_ab} and \code{CorAb_aB})
between iterations is smaller than this value, the iteration will stop. Default is 1e-6.}

\item{dat}{Specifies which iteration results to return:
\itemize{
\item \code{dat = 1} returns only the final results of the last iteration for each population.
\item \code{dat = 2} returns the full iteration history for each population.
Default is \code{dat = 1}.
}}
}
\value{
A data frame containing the haplotype frequencies (\code{AB}, \code{Ab}, \code{aB}, \code{ab}) for each population.
The output also includes the iteration number, population name, total sample size (\code{N}), and the correction factors
(\code{CorAB_ab}, \code{CorAb_aB}) for each iteration.
}
\description{
This function computes the haplotype frequencies for a biallelic genetic marker (AB, Ab, aB, ab) in multiple populations
based on a genotype distribution. The method accounts for genotype counts and utilizes an iterative approach to
estimate the haplotype frequencies. The results can be returned as the final iteration or with the full iteration history
for each population. This function is designed to work with a data frame that contains genotype counts for each population.
}
\details{
The function uses a correction approach to estimate haplotype frequencies.
The iterative process aims to converge by adjusting the correction factors (\code{CorAB_ab} and \code{CorAb_aB}) until their change
between iterations is smaller than the specified tolerance (\code{tol}). The iteration history can be useful for assessing convergence
behavior or troubleshooting the process.
}
\examples{
# Create a data frame with genotype counts for three populations
DicGtps <- data.frame(
  Pop = c("Pop 1", "Pop 2", "Pop 3"),
  AABB = c(5, 3, 4),
  AABb = c(3, 2, 5),
  AAbb = c(4, 2, 5),
  AaBB = c(4, 6, 3),
  AaBb = c(10, 8, 12),
  Aabb = c(6, 5, 7),
  aaBB = c(3, 4, 2),
  aaBb = c(2, 1, 3),
  aabb = c(3, 2, 4)
)

# Calculate haplotype frequencies for the data frame
Diloc_Freq_Hd(DicGtps, max_iterations = 10, tol = 1e-6, dat = 1)

}
