% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomMating.r
\name{randomMating}
\alias{randomMating}
\title{random mating at 2 loci that have either susceptible or resistant alleles}
\usage{
randomMating(G)
}
\arguments{
\item{G}{array with frequencies of gametes}
}
\value{
array with frequencies of genotypes in expanded format
}
\description{
constructs genotypes given gametes
returns array of genotype frequencies in expanded format
i.e. the 16 exhaustive combinations of 2 alleles at 2 loci (2x2x2x2).
These can be converted to the collapsed 10 genotype format (8+cis&trans for the double heterozygotes),
by \code{\link{genotypesLong2Short}}
This is refactored from code in runModel()
}
\examples{
#randomMating()
}

