% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomMating.r
\name{randomMating}
\alias{randomMating}
\title{random mating}
\usage{
randomMating(f, G)
}
\arguments{
\item{f}{array with frequencies of genotypes in popn}

\item{G}{array with frequencies of gametes}
}
\value{
array with frequencies of genotypes
}
\description{
constructs genotypes given gametes
returns array of genotype frequencies
This is refactored from code in runModel()
}
\examples{
#randomMating()
}

