% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resistanceMaster.r
\name{resistanceMaster}
\alias{resistanceMaster}
\title{master function to setup and run resistance model scenarios}
\usage{
resistanceMaster(params.csv = TRUE, inFile = "input.parameters.csv",
  calibration = 100, produce.plots = TRUE, savePlots = FALSE)
}
\arguments{
\item{params.csv}{whether parameters are defined in an input file TRUE/FALSE}

\item{inFile}{parameters filename}

\item{calibration}{one of a limited set of integers effecting how scenarios are run}

\item{produce.plots}{whether to produce plots}

\item{savePlots}{whether to save plots to hardcoded filenames}
}
\value{
nothing yet
}
\description{
May restructure later.
developed from malaria_code_andy.r
}

