% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/resistSimple.r
\name{resistSimple}
\alias{resistSimple}
\title{to run one simple scenario without an input file, experimental}
\usage{
resistSimple(...)
}
\arguments{
\item{...}{extra arguments passed to setInputOneScenario()}
}
\value{
a list of 3 lists of one or more scenarios: results, genotype and fitness. e.g. listOut$results[1] gives a results matrix for the first scenario
}
\description{
May restructure later.
}

