% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Single_loc_freq.R
\name{Single_loc_freq}
\alias{Single_loc_freq}
\title{Calcular frecuencias genotípicas y alélicas}
\usage{
Single_loc_freq(freq_matrix)
}
\arguments{
\item{freq_matrix}{Un data frame con las siguientes columnas:
\describe{
\item{Pop}{Un identificador para cada población (e.g., "Pop1", "Pop2").}
\item{aa}{El conteo de individuos con el genotipo aa.}
\item{Aa}{El conteo de individuos con el genotipo Aa.}
\item{AA}{El conteo de individuos con el genotipo AA.}
}}
}
\value{
Un data frame con las siguientes columnas:
\describe{
\item{Population}{El identificador de la población.}
\item{n}{El número total de individuos considerados en los cálculos.}
\item{aa}{La frecuencia genotípica de individuos con el genotipo aa.}
\item{Aa}{La frecuencia genotípica de individuos con el genotipo Aa.}
\item{AA}{La frecuencia genotípica de individuos con el genotipo AA.}
\item{Freq_A}{La frecuencia alélica del alelo A.}
\item{Freq_a}{La frecuencia alélica del alelo a.}
}
}
\description{
Esta función calcula las frecuencias genotípicas y alélicas a partir de una matriz que contiene los conteos de genotipos (AA, Aa, aa) por población.
}
\examples{
# Ejemplo: Matriz de frecuencias genotípicas para una población
freq_matrix <- data.frame(
  Pop = c("Pop1"),
  aa = c(2),
  Aa = c(9),
  AA = c(1))

# Calcular frecuencias
Single_loc_freq(freq_matrix)

# Ejemplo: Matriz de frecuencias genotípicas para multiples poblaciones


  df2 <- Import_single_loc(Loc1Pops3)

  Single_loc_freq(df2)
}
